Added large file (64-bit) support (STR #541) and job abort code to lpd backend.
backend/lpd.c:
    - Add abort_job global to track when SIGTERM is received.
    - lpd_command(), lpd_write(): Check for abort_job.
    - lpd_queue(): Increase buffer size to 64k, and use off_t and
      CUPS_LLFMT/CUPS_LLCAST to handle large files.
    - sigterm_handler(): Set abort_job instead of exiting.

berkeley/lpq.c:
    - show_jobs(): Update job-k-octets code to handle large
      files.

config.h.in:
    - Add "long long" configure definitions.

config-scripts/cups-compiler.m4:
    - Fix comment header.

config-scripts/cups-largefile.m4:
    - Added.

configure.in:
    - Include cups-largefile.m4.

cups/http.c:
    - httpGetLength(): Call httpGetLength2() and return
      _data_remaining.
    - httpGetLength2(): Added.
    - httpRead(): Use stroll() to get chunk length, and mirror
      data_remaining to _data_remaining.
    - httpSetLength(): Added.
    - httpUpdate(): Use httpGetLength2().
    - http_send(): Use httpGetLength2().

cups/http.h:
    - Add new off_t data_remaining member to http_t.
    - Rename int data_remaining member in http_t to
      _data_remaining.
    - Add prototypes for httpGetLength2() and httpSetLength().

cups/ipp.c:
    - ippRead(): Update debug printf for large file sizes.
    - ipp_read_http(): Mirror data_remaining to _data_remaining.

cups/testhttp.c:
    - main(): Use off_t, CUPS_LLFMT/CUPS_LLCAST, and
      httpGetLength2().

cups/util.c:
    - cupsDoFileRequest(): Use off_t, CUPS_LLFMT/CUPS_LLCAST, and
      httpGetLength2(), increase buffer size to 64k.

cups.spec:
    - Add missing cancel man page, and move backend and filter
      backends to -devel package.

Makedefs.in:
    - Add @LARGEFILE@ to CFLAGS and CXXFLAGS.

pdftops/SecurityHandler.cxx:
    - SecurityHandler::make(): #ifdef xsh variable definition.

scheduler/client.c:
    - cupsdReadClient(): Intialize _data_remaining member, use
      off_t and CUPS_LLFMT/CUPS_LLCAST to support large files.
    - cupsdSendFile(): Use CUPS_LLFMT/CUPS_LLCAST to support
      large files.
    - check_if_modified(): Use strtoll() and
      CUPS_LLFMT/CUPS_LLCAST to support large files.
    - pipe_command(): Use CUPS_LLFMT/CUPS_LLCAST to support large
      files.

scheduler/client.h:
    - Change bytes member in cupsd_client_t to off_t.

scheduler/cups-lpd.c:
    - send_state(): Change job-k-octets handling to support large
      files.

scheduler/ipp.c:
    - cupsdProcessIPPRequest(): Update Content-Length code to
      support large files and mirror data_remaining to
      _data_remaining.

scheduler/log.c:
    - cupsdLogRequest(): Use CUPS_LLFMT/CUPS_LLCAST to log large
      file sizes.

systemv/lpstat.c:
    - show_jobs(): Change job-k-octets handling to support large
      files.


git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@4730 7a7537e8-13f0-0310-91df-b6672ffda945

+++ scheduler/client.h
@@ -34,7 +34,7 @@ struct cupsd_client_s
   cupsd_location_t	*best;		/* Best match for AAA */
   time_t		start;		/* Request start time */
   http_state_t		operation;	/* Request operation */
-  int			bytes;		/* Bytes transferred for this request */
+  off_t			bytes;		/* Bytes transferred for this request */
   char			username[33],	/* Username from Authorization: line */
 			password[33],	/* Password from Authorization: line */
 			uri[HTTP_MAX_URI],
